fix: changed the floats to int32_t

+++ RTD_firmware/Core/Inc/ADS1248.h
@@ -21,9 +21,9 @@ void ADS124X_init(SPI_HandleTypeDef *hspi,
         GPIO_TypeDef *GPIO_START, uint16_t PIN_START,
         GPIO_TypeDef *GPIO_CS, uint16_t PIN_CS);
 
-float Temperature(SPI_HandleTypeDef *hspi,
+int32_t Temperature(SPI_HandleTypeDef *hspi,
         GPIO_TypeDef *GPIO_CS, uint16_t PIN_CS);
 
-float RTD_Converter(int32_t result);
+int32_t RTD_Converter(int32_t result);
 
 #endif /* INC_ADS1248_H_ */
